requirements.txt
# Core dependencies
python-dotenv==1.0.0
psycopg2-binary==2.9.9

# pgvector support
pgvector==0.2.4

# LangChain ecosystem
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2

# Groq AI
groq==0.4.1

# Embeddings & utilities
sentence-transformers==2.2.2
tiktoken==0.5.2
numpy==1.24.3

# API & Web
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Data processing
pandas==2.1.4
sqlalchemy==2.0.25